+++ doc/changelog.md
@@ -2,6 +2,7 @@
 ## Breaking changes
 - PathSet class no longer exists, replaced by a function of the same name that creates a new instance only if needed. If you have code that subclasses PathSet change it to subclass BasePathSet. 
 - BasePathSet._resolveUnderlyingDepenencies() now returns a generator of (path, pathset) instead of a list of [path]. This only affects users with a custom subclass of BasePathSet with an override of this method (and does not affect you if you used DerivedPathSet). 
+- Target priority can no longer be set to a negative number; 0.0 is the minimum.
 
 ## Deprecation
 None